refactor(lazer): move ed25519 instruction code to solana contract package
