+++ lazer/contracts/solana/package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "@coral-xyz/anchor": "^0.30.1",
     "@pythnetwork/pyth-lazer-sdk": "^0.3.1",
-    "@solana/web3.js": "^1.98.0"
+    "@solana/web3.js": "^1.98.0",
+    "@solana/buffer-layout": "^4.0.1"
   },
   "devDependencies": {
     "@types/bn.js": "^5.1.0",
