+++ lazer/sdk/js/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@pythnetwork/pyth-lazer-sdk",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "Pyth Lazer SDK",
   "publishConfig": {
     "access": "public"
@@ -59,8 +59,6 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@isaacs/ttlcache": "^1.4.1",
-    "@solana/buffer-layout": "^4.0.1",
-    "@solana/web3.js": "^1.98.0",
     "isomorphic-ws": "^5.0.0",
     "ts-log": "^2.2.7",
     "ws": "^8.18.0"
