+++ lazer/sdk/js/src/constants.ts
@@ -1,8 +1,4 @@
-import { PublicKey } from "@solana/web3.js";
-
-export const SOLANA_LAZER_PROGRAM_ID = new PublicKey(
-  "pytd2yyk641x7ak7mkaasSJVXh6YYZnC7wTmtgAyxPt",
-);
-export const SOLANA_STORAGE_ID = new PublicKey(
-  "3rdJbqfnagQ4yx9HXJViD4zc4xpiSqmFsKpPuSCQVyQL",
-);
+export const SOLANA_LAZER_PROGRAM_ID =
+  "pytd2yyk641x7ak7mkaasSJVXh6YYZnC7wTmtgAyxPt";
+export const SOLANA_LAZER_STORAGE_ID =
+  "3rdJbqfnagQ4yx9HXJViD4zc4xpiSqmFsKpPuSCQVyQL";
