+++ lazer/sdk/js/src/index.ts
@@ -1,4 +1,3 @@
 export * from "./client.js";
 export * from "./protocol.js";
-export * from "./ed25519.js";
 export * from "./constants.js";
